extending pluggable apps. create new app called generic app manager. it will handle dependency (generic) between multiple apps.
now I have shifted TagInline from 'store admin' to 'new_app admin.'

so there's no dependency between store and tags app
