+++ generic_app_manager/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class GenericAppManagerConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'generic_app_manager'
