+++ store/admin.py
@@ -1,11 +1,9 @@
 from django.contrib import admin, messages
-from django.contrib.contenttypes.admin import GenericTabularInline
 from django.db.models import Count
 from django.urls import reverse
 from django.utils.html import format_html, urlencode
 
-from .models import Collection, Customer, Product, Promotion, Order, OrderItem
-from tags.models import TaggedItem
+from .models import Collection, Customer, Product, Order, OrderItem
 
 
 # Register your models here.
@@ -44,13 +42,6 @@ class CollectionAdmin(admin.ModelAdmin):
         return super().get_queryset(request).annotate(products_count=Count('product'))
 
 
-class TagInline(GenericTabularInline):
-    autocomplete_fields = ['tag']
-    model = TaggedItem
-    min_num = 1
-    max_num = 10
-    extra = 0
-
 
 @admin.register(Product)
 class ProductAdmin(admin.ModelAdmin):
@@ -59,7 +50,6 @@ class ProductAdmin(admin.ModelAdmin):
         'slug': ['title']
     }  # this will auto-populate slug field with title
     actions = ['clear_inventory']
-    inlines = [TagInline]
     list_display = ['title', 'price', 'inventory_status', 'collection']
     list_editable = ['price']
     list_per_page = 10
